+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
+## Unreleased
+
+### Added
+
+- Plugins can be deprecated now (CMEM-6706)
+
 ## [4.16.1] 2026-02-20
 
 ### Fixed